Moved shared script to utils
